+++ changelog.md
@@ -3,6 +3,7 @@
     - Textures kindly provided by AriaElidove, thank you!
 - Remove raw ore blocks from the list of blocks dug by the grinding disk  (voidsong-dragonfly)
 - Change workbench to be able to handle blueprints with more than 9 recipes (BluSunrize)
+- Fix villagers being unable to interact with IE doors (BluSunrize)
 
 ##### Version 1.20.4-11.4.3-178
 - Change scaffolding blocks to have more descriptive names (AriaElidove)